requirements.txt
openai
faster-whisper
sounddevice
soundfile
python-dotenv
numpy
PyQt6
pyinstaller
